Support a barebones "city screen" view
This is triggered when building a city and when clicking "zoom to city"

For now I've just added the "exit" button, but hope to add tile yields soon.

I also fixed an issue with the advisor and input handling - previously if you exited the advisor via escape the game would ask if you wanted to exit. Now it works properly.