Display tile yields when zooming to a city
This allows us to see what tiles are being worked, and also makes it clearer that we are currently giving the wrong yield for city centers.

In order for this to render properly I had to fix some issues with how the map view layers worked. Previously all the layers we part of the same `Node2D`, and most things were draw as textures. However, cities and units are child nodes, and child nodes are drawn on top of textures. This usually isn't a problem, but it meant that drawing the city center tile yield was hidden by the city itself. This also is the reason health bars were partially hidden by the city - the health bar is a texture, the city is a node.

To fix this I grouped the layers into separate nodes, so that within each node the layer ordering was simple, and then the node ordering took effect after that. As a result health bars now render properly, and we can see the city center yield.