Simplify the logic for the lower right info box so it stays up to date
For the city scene, where we show the name, what's produced, and how many turns are left until production/growth, we just grab information from the game data when drawn. We can do that for the lower right info box as well, ensuring that things are up to day whenever something happens, like a road being finished or a citizen being moved to a different tile.

This allows us to get rid of some unneeded event logic.