Fix black spots on loaded PCX textures
The issue fixed by this commit was introduced in PR #504. It was
caused by the addition of a new color index to the set of transparent
colors. This new index was required to correctly display border
textures, however, when applied to other textures it produced
black spots in areas that should have remained opaque.

This commit modifies the methods for loading PCX textures, so they can
accept a custom set of transparent color indexes. To improve
readability, two new structs are introduced:

- ColorOptions: Encapsulates transparency and shadows settings.
- CropRegion: Defines a cropped section of a texture.