Load requirements for building and unit production
This commits implement loading of several production
requirements. For both buildings and units, it loads the required
technology. For buildings it also loads the required building (like a
Library for a University) and the wonder status.

Additionally, this commit refactors the ToGameData method of the
SaveGame class, splitting it into multiple methods for better
readability.

With these changes we are getting closer to creating a proper method
for calculating a list of producible units and buildings, although we
still need to load data on unique units and unit obsolescence.