Rewrite ExplorerAI and Return an enum from `PlayTurn`, not a bool
The previous return value was confusing and conflated "I'm done with my goal" and "I hit an error trying to reach my goal". This resulted in stack overflows when using the automated explore functionality on a human unit, resulted in "goto" being used in the settler ai, and some other oddities that would pop up when running in observer mode.

The new enum allows better signaling of how the AI is progressing and better error handling.

I also consolidated the logic for "move the AI unit along a path", since it was repeated in a couple of places. The annoying C7Engine/C7GameData split required that I put this in an extension class rather than being part of the UnitAI class as a protected method, like I'd prefer.

This work also revealed that the explorer ai was overly complicated and different from the other unit AIs in that it didn't fit the pattern of the AI data being a plan, and PlayTurn executing the plan. To solve that I rewrote it, and the resulting behavior is much nicer:
 - we no longer get big clumps of explorers heading for the same tile at different speeds
 - debugging what it is doing is easier
 - we have to check far fewer tiles because we take advantage of the border tiles we track

Lastly, I had the AI evaluation skip fortified units, which sped things up a bunch and stopped log spam from units repeatedly fortifying themselves as city defenders.

Continued work on rewriting ai pieces