Stop the AI from repeatedly attempting to walk across oceans
While working to speed up the AI turns after turn 100, I found that the AI was repeatedly attempting to send settlers to other continents by foot. Attempting this pathing is the worst case scenario for the pathing algorithms, as it has to check every tile on the continent to discover there is no route.

To solve this, we just add an identifier to each land mass, to make it trivial to exit early if no land route is possible. I'm fairly certain the base game did something similar, since the base game had an existing "continents" field for each tile.