Implement war declarations via the diplomacy UI
We can now declar war on any civs that we have met, and they will properly be at war with us.

This doesn't implement peace treaties yet, doesn't handle the actual flc files for leaders (and just uses the pcx of the head), doesn't handle leader emotions (just hardcodes cautious), and doesn't handle refusing to answer communication. Howver, this PR is already pretty big and adds a pretty important bit of functionality.

The pedia icons logic is stuff we'll need eventually, so even though we might want the real flc files in the future, this is a good standin and can be reused for the foreign advisor screen.