Correct the calculation for rank distances
The previous formula only worked up to 99k culture, and as I intend to use this function for corruption-related calculations, it needs to be accurate to further distances.

The piece I was missing from the earlier version of the formula was that it is calculated on the NW/SE NE/SW grid, not the x y grid. Once we transform the x y grid to the NW/SE NE/SW grid (by rotating coordinates 45 deg) the formula from [the corruption post](https://forums.civfanatics.com/threads/everything-about-corruption-c3c-edition.76619/) is simple.