Allow changing governments!
Once the despotism penalty PR is in, this will allow getting out of despotism. This PR handles hooking up the button, figuring out how many turns of anarchy there are, prompting the user to select a government when anarchy is over, and actually using the new government.

It doesn't handle:
 - AI government switching
 - Religious civs
 - Anarchy's negative effects
 - Prompting the user that a government change is possible once researched