Handle happiness levels from luxury spending and default unhappiness
This is just the start to the rather complex world of happiness calculations, but it gives us a place to start with many other fancier things:
 - Making hooking up luxuries do something
 - Making marketplaces functional
 - Being able to add civil disorder
 - Making the domestic advisor show city populations and moods
 - Implementing pop rushing/drafting and the penalties
 - Implementing entertainers and other specialists

Now when a city in the default game hits size 3, one of the citizens will be unhappy, since only two citizens are born content on regent difficulty. We can fix this with the luxury slider (note the 1 gpt of luxury spending in the second screenshot).