Introduce a custom C7-specific terrain pack
This includes custom terrain, custom mines, custom resources, custom irrigation, and custom forts/barb camps.

It does not yet include fog of war, roads, or railroads.

Since I have very little artistic talent, the bulk of this was done via the following steps:

1. Using [WildWeazel's terrain templates](https://forums.civfanatics.com/threads/templates-for-modding.130897/post-3181736) and picking the terrain color.

2. Removing the pink and replacing it with transparency.

3. In paint.net, duplicating the layer.

4. In one layer copy, applying the "Dent" effect, adding noise, applying the "oil painting" effect, and adding noise again. This shinks the tiles somewhat, so leaving the base layer unchanged is important.

5. Flattening the labels and exporting as a png. I've included the raw `.pdn` files to make this easier to edit if we want to change it in the future.

The resources are the same as what we had before, but because it is now a `.png` file, I was able to use a better resizing algorithm that looks better, and real transparency.

----

To make using our own art configurable, I made "Ctrl+Alt+Shift+T" the shortcut for toggling the new art.