Add building-level mood modifications in a particular city
This only applies to city-level effects, not global or continent-level effects, but means that temples will start to be functional.

Interestingly there's an "unhappiness" modifier for buildings available in the editor, though the base game doesn't actually use this. I did some testing with a modified scenario and found that it moves happy faces back to content.