Implement hurrying production
Right now this is only via the right click menu, not via the city screen, but it supports pop rushing and gold rushing, depending on the form of government.

I haven't yet added the happiness penalty for pop rushing; I'll do that in a following PR since it'll require thinking about negative happiness and weird things.