Split some noise generation out to its own class
And add a test that generates images so the noise can be visualized. I think adding a dependency here to allow this is worth it, as debugging some of the terrain generation is extremely hard without visualizing it, and being able to visualize it without launching the game saves a lot of time.