Make civ3 graphics the default and add a super basic toggle button
This should help unstick some of the discussions around AI art by making it clearly an opt-in situation that only exists if a player doesn't want to use civ3 graphics and isn't using a mod that provides a replacement.

This should probably live in a preferences screen someday, but this is good enough for now.