Support great wonders that add buildings to other cities
I originally implemented this via https://github.com/C7-Game/Prototype/pull/730, which added the buildings directly to the city's list of buildings when the wonder was created or the building was created, and removed them if the wonder was destroyed. However, this gets much more complicated once wonders can go obsolete, and especially if there was a scenario where two wonders provided the same buildings. As a result I reworked the PR to instead have us get any relevant wonder-supplied buildings when we iterate over the list of buildings, which works much nicer.