Optimize the calculation of trade networks
Rather than recomputing the pathing on the fly every time we need a city's luxury count, we can compute the whole network and cache it, and only invalidate it when roads are built (and eventually pillaged) and when cities are built and destroyed.

This cuts the runtime of the load conquest scenarios test from ~53 seconds to ~16 seconds on my laptop.