Use a single trade network object for all players
This ensures we always recalculate all the trade networks whenever we invalidate it. Any time one player's trade network would be invalidated we need to do so for all of them - player A's border changing could influence what resources player B has. Player A building a city could impact the transitive road network player B depends on. Ditto with player A building a road, etc etc.

Doing this required pushing some access to the `EngineStorage.gameData` singleton up a few levels because we indirectly look up resources when loading scenarios. This happens when doing the initial citizen assignment, since we need to know how happy the city is to properly place the citizens. In just about every case pushing the call up improved the code, either fixing a latent issue where the singleton was accessed without proper locking, or making it easier to track the lock.

In [this review comment](https://github.com/C7-Game/Prototype/pull/745#issuecomment-2952690460) we discussed ignoring resource owners until it is time to look up resources, but I went with this route because I realized we always need to recalculate all the players' trade routes. So we don't need to ignore resource owners.

I also made the trade network more of a class, which more easily allows for checking whether the player knows about a resource. However this time I did it at the time of checking for resource access, not when building the trade network, so there's no need to invalidate when a tech is discovered (until techs like navigation).