Add custom unit action buttons
The icon art was generated using imagefx with prompts along these lines:

```
A set of 8 circular, icons for a strategy game. The icons should be a solid white circle. The symbols on the icons should be a dark brownish-green color with a slightly worn, hand-painted style. The icons should represent the following actions: 
1. A machete in front of two hanging curly jungle vines
2. two thick, parallel vertical lines in a dark, earthy brown symbolizing a road. 
3. A toxic waste indicator
4. Two crossed rifles
5. A parabola with a cannon ball at the end
6. The silhouette of a lightning bolt
8. A fur trading colony
The symbols should be presented in a row on a white background
```

I then edited them in paint.net (stripping backgrounds, resizing, making lines thicker/thinner).

The rollover/highlight variants were done using the "BoltBait's Color Balance" adjustment tool, taking the RGB values of the orange/blue I've used for other buttons (2ED3FA, ED730E) and getting them on a 0-100 scale for the tool.