Fix graphical issues with CityLayer
Previously, the city graphics disappeared when wrapping around the
map. This was caused by setting a city scene position only on scene
creation, while it required updates on map wrapping. This commit fixes
it by updating a city scene position on each frame.

This commit also fixes a bug that caused the city graphics to be
rendered in the wrong size around the world wrap line. It was caused
by setting the city graphics' OffsetLeft to a negative value. It was
fixed by updating the Position property instead, which correctly
handles negative values.

Also, this commit introduces a performance improvement by loading the
disorder animation in the static CityScene constructor, instead of
loading it on every scene initialization. On my end, it noticeably
improved the frame rate.