Refactor: improve game creation
This commits improves mechanism for game creation.

Previously, some logic for initializing the game state was run as part
of the save loading class, before the EngineStorage was
initialized. As a result, some methods had accepted GameData object as
an argument, instead of reading it from the EngineStorage, as is
commonly done in the codebase.

In my view, this was problematic, since it introduced a split in
codebase and required adjusting the code based on whether it would be
run during game creation. This commit addresses this issue by
introducing an action field "onGameCreation" to GameData. Now, the
SaveGame class only binds the initializing logic to this action, which
is executed in the CreateGame class after the EngineStorage is
properly initialized.

Methods that accept GameData as argument still exist in the
project. Some of them relate to the game creation mechanism, and some
of them were written this way because previously it wasn't possible to
access EngineStorage from the C7GameData namespace. I think we should
gradually refactor them out in favor of accessing GameData through the
EngineStorage.