Terraform AI scoring
This commit updates the AI for choosing the best worker
action (terraform) for a tile.

Previously, the AI scored terraforms by making a copy of a tile,
applying a terraform to it and comparing yields of the original tile
and its copy. This system had problems. If the terraform had an effect
on the world outside of a tile (for example, forest chopping granting
shields to the nearest city), these changes were mistakenly applied
during the AI score calculations. There was also no way to account for
such effects in AI calculations.

This commit fixes that by extending the Terraform class with a new Lua
"hook" for AI score calculations. This is a function bound to a
terraform that returns its AI points. This way, the effect of the
terraform is run only on its completion. This also allows implementing
custom AI for new terraforms without modifying the game core.